+++ CheckmarxPythonSDK/CxOne/httpRequests.py
@@ -24,15 +24,19 @@ def get_data_from_config():
             "client_id": "ast-app",
             "refresh_token": config.get("refresh_token"),
         }
-    return server_url, token_url, timeout, verify_ssl_cert, cert, token_req_data
+    proxies = {
+        "http": config.get("proxy"),
+        "https": config.get("proxy"),
+    }
+    return server_url, token_url, timeout, verify_ssl_cert, cert, token_req_data, proxies
 
 
 get, post, put, patch, delete, head = build_request_funcs(get_data_from_config)
 
 headers = {
-    "user-agent":
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) "
-        "Chrome/106.0.0.0 Safari/537.36"
+    # "user-agent":
+    #     "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) "
+    #     "Chrome/106.0.0.0 Safari/537.36"
 }