+++ CheckmarxPythonSDK/CxScaApiSDK/config.py
@@ -12,6 +12,7 @@ config_default = {
     "timeout": 60,
     "verify": False,
     "cert": None,
+    "proxy": None,
 }
 
 config = get_config(config_default=config_default, section="CxSCA", prefix="cxsca_")