+++ tests/CxOne/test_sast_results_summary_api.py
@@ -1,17 +1,11 @@
 from CheckmarxPythonSDK.CxOne import (
-    get_summary_for_many_scans,
     get_sast_aggregate_results,
 )
 
 
-def test_get_summary_for_many_scans():
-    response = get_summary_for_many_scans(scan_ids=["353d9a5d-cf52-4384-8516-e31d7447ead1"])
-    assert response is not None
-
-
 def test_get_sast_aggregate_results():
     response = get_sast_aggregate_results(
-        scan_id="353d9a5d-cf52-4384-8516-e31d7447ead1",
+        scan_id="09ad7faf-74e5-415b-b81a-f4f209b736a4",
         group_by_field=["LANGUAGE"],
     )
     assert response is not None